Added various properties and helper methods to parser classes.

+++ d11-boot-parser/src/main/java/org/d11/boot/parser/model/GoalData.java
@@ -19,7 +19,7 @@ public class GoalData extends ParserModel {
     /**
      * Whoscored id of the team that scored the goal.
      */
-    private Long teamWhoscoredId;
+    private Integer teamWhoscoredId;
     /**
      * Name of the team that scored the goal.
      */
